pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ethan Brown <ethan@ewbrowntech.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
slowapi = "^0.1.9"
ffmpeg-python = "^0.2.0"
python-multipart = "^0.0.9"
docker = "^7.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
ffmpeg-python = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
